LayoutTests/platform/mac-monterey/http/tests/security/xss-DENIED-xsl-external-entity-xslt-docloader-expected.txt

CONSOLE MESSAGE: Unsafe attempt to load URL 127.0.0.1:8000. Domains, protocols and ports must match.

CONSOLE MESSAGE: Failure to process entity passwd

CONSOLE MESSAGE: Entity 'passwd' not defined

CONSOLE MESSAGE: Unsafe attempt to load URL 127.0.0.1:8000. Domains, protocols and ports must match.

CONSOLE MESSAGE: Failure to process entity hosts

CONSOLE MESSAGE: Entity 'hosts' not defined

CONSOLE MESSAGE: Unsafe attempt to load URL 127.0.0.1:8000. Domains, protocols and ports must match.

CONSOLE MESSAGE: Failure to process entity group

CONSOLE MESSAGE: Entity 'group' not defined

ALERT: document()
